rename test fixture to building request class and add object parameter to test method
