+++ test/blueprints/Do.Test.Blueprints.Service/Business/MethodSamples.cs
@@ -38,8 +38,8 @@ public class MethodSamples(ILogger<MethodSamples> _logger)
         _logger.LogInformation($"{nameof(VoidAsync)} was called");
     }
 
-    public object RequestClass(string @string, int @int) =>
-        new { @string, @int };
+    public object RequestClass(string @string, int @int, object @object) =>
+        new { @string, @int, @object };
 
     public object Object(object @object) =>
         @object;
